+++ src/aqua_diagnostics/ensemble/cli_timeseries_ensemble.py
@@ -4,265 +4,269 @@ Command-line interface for ensemble global time series diagnostic.
 
 This CLI allows to plot ensemle of global timeseries of a variable
 defined in a yaml configuration file for multiple models.
+
+NOTE: Since the reference data is not in the catalog the data is loaded from the path
+      Once the reference data is uploaded in the catalog, line 170-192 can be un-commented 
+      and line 194-217 can be removed/commented.
 """
+
 import argparse
-import os
 import sys
-import gc
+
 import xarray as xr
-from dask.distributed import Client, LocalCluster
-import numpy as np
-import pandas as pd
-from aqua.util import load_yaml, get_arg, ConfigPath
-from aqua.exceptions import NotEnoughDataError, NoDataError, NoObservationError
+from aqua.util import get_arg
 from aqua.logger import log_configure
-from aqua import Reader
+from aqua.version import __version__ as aqua_version
+from aqua.diagnostics.core import template_parse_arguments, open_cluster, close_cluster
+from aqua.diagnostics.core import load_diagnostic_config, merge_config_args
+
+from aqua.diagnostics import retrieve_merge_ensemble_data
 from aqua.diagnostics import EnsembleTimeseries
+from aqua.diagnostics import PlotEnsembleTimeseries
+
 
 def parse_arguments(args):
-    """Parse command line arguments."""
+    """Parse command-line arguments for EnsembleTimeseries diagnostic.
 
-    parser = argparse.ArgumentParser(description="Global time series CLI")
-
-    parser.add_argument("-c", "--config",
-                        type=str, required=False,
-                        help="yaml configuration file")
-    parser.add_argument('-n', '--nworkers', type=int,
-                        help='number of dask distributed workers')
-    parser.add_argument("--loglevel", "-l", type=str,
-                        required=False, help="loglevel")
-
-    # These will override the first one in the config file if provided
-    # parser.add_argument("--catalog", type=str,
-    #                    required=False, help="catalog name")
-    parser.add_argument("--model", type=str,
-                        required=False, help="model name")
-    parser.add_argument("--exp", type=str,
-                        required=False, help="experiment name")
-    parser.add_argument("--source", type=str,
-                        required=False, help="source name")
-    parser.add_argument("--outputdir", type=str,
-                        required=False, help="output directory")
+    Args:
+        args (list): list of command-line arguments to parse.
+    """
+    parser = argparse.ArgumentParser(description="EnsembleTimeseries CLI")
+    parser = template_parse_arguments(parser)
     return parser.parse_args(args)
 
 
-def get_plot_options(config: dict = None, variable: str = None):
-    """
-    Extracts timeseries plot options from a configuration dictionary.
-
-    This function retrieves a set of parameters related to timeseries plotting from the 
-    `timeseries_plot_params` key of the provided config file.
-
-    Args:
-        config (config file): Settings are defined in the config file 
-            which is load by the load_yaml function. 
-            It is expected to include the key `timeseries_plot_params` with 
-            sub-keys for various plotting parameters. Defaults to None.
-        variable (str): A variable name (not used in the current implementation, 
-            but reserved for future use). Defaults to None.
-            
-    Returns:
-        tuple: A tuple containing the following elements extracted from the 
-        `timeseries_plot_params` key in the configuration:
-            - startdate (any): The start date for the timeseries plot (default: None if not found).
-            - enddate (any): The end date for the timeseries plot (default: None).
-            - plot_std (any): Flag or settings for plotting standard deviations (default: None).
-            - plot_ensemble_members (any): Flag or settings for plotting ensemble members (default: None).
-            - ensemble_label (any): Label for ensemble data (default: None).
-            - figure_size (any): Size of the figure (default: None).
-            - ref_label (any): Label for the reference data (default: None).
-            - label_ncol (any): Number of columns for the plot legend labels (default: None).
-            - label_size (any): Font size of the labels (default: None).
-            - pdf_save (any): Whether to save the plot as a PDF (default: None).
-            - units (any): Units for the data being plotted (default: None).
-    """
-    startdate = config["timeseries_plot_params"].get("startdate", None)
-    enddate = config["timeseries_plot_params"].get("enddate", None)
-    plot_std = config["timeseries_plot_params"].get("plot_std", None)
-    plot_ensemble_members = config["timeseries_plot_params"].get("plot_ensemble_members", None)
-    ensemble_label = config["timeseries_plot_params"].get("ensemble_label", None)
-    figure_size = config["timeseries_plot_params"].get("figure_size", None)
-    ref_label = config["timeseries_plot_params"].get("ref_label", None)
-    label_ncol = config["timeseries_plot_params"].get("label_ncol", None)
-    label_size = config["timeseries_plot_params"].get("label_size", None)
-    pdf_save = config["timeseries_plot_params"].get("pdf_save", None)
-    units = config["timeseries_plot_params"].get("units", None)
-    return startdate, enddate, plot_std, plot_ensemble_members, ensemble_label, figure_size, ref_label, label_ncol, label_size, pdf_save, units
-
-
-def retrieve_data(variable=None, models=None, exps=None, sources=None, startdate=None, enddate=None, ens_dim="Ensembles"):
-    """
-    Retrieves, merges, and slices datasets based on specified models, experiments, 
-    sources, and time boundaries.
-
-    This function reads data for a given variable (`variable`) from multiple models, experiments, 
-    and sources, combines the datasets along the specified ensemble dimension, and slices 
-    the merged dataset to the given start and end dates.
-
-    Args:
-        variable (str): The variable to retrieve data for. Defaults to None.
-        models (list): A list of model names. Each model corresponds to an 
-            experiment and source in the `exps` and `sources` lists, respectively. 
-            Defaults to None.
-        exps (list): A list of experiment names. Each experiment corresponds 
-            to a model and source in the `models` and `sources` lists, respectively. 
-            Defaults to None.
-        sources (list): A list of data source names. Each source corresponds 
-            to a model and experiment in the `models` and `exps` lists, respectively. 
-            Defaults to None.
-        startdate (str or datetime): The start date for slicing the merged dataset. 
-            If None, the maximum start date across all datasets is used. Defaults to None.
-        enddate (str or datetime): The end date for slicing the merged dataset. 
-            If None, the minimum end date across all datasets is used. Defaults to None.
-        ens_dim (str, optional): The name of the dimension along which the datasets are 
-            concatenated. Defaults to "Ensembles".
-
-    Returns:
-        tuple:
-            - startdate (datetime): The resolved start date for the sliced dataset.
-            - enddate (datetime): The resolved end date for the sliced dataset.
-            - xarray.Dataset: The merged dataset containing data from all specified models, 
-              experiments, and sources, concatenated along `ens_dim` and sliced 
-              to the time range `[startdate, enddate]`.
-    """
-    dataset_list = []
-    startdate_list = []
-    enddate_list = []
-    if models is None or exps is None or sources is None:
-        raise NoDataError("No models, exps or sources provided")
-    else:
-        for i, model in enumerate(models):
-            reader = Reader(model=model, exp=exps[i], source=sources[i], areas=False, startdate=startdate, enddate=enddate)
-            data = reader.retrieve(var=variable)
-            dataset_list.append(data)
-            startdate_list.append(data.time[0].values)
-            enddate_list.append(data.time[-1].values)
-    merged_dataset = xr.concat(dataset_list, ens_dim)
-    
-    if startdate != None:
-        istartdate = max(startdate_list)
-        startdate = pd.to_datetime(startdate) 
-        startdate = max(istartdate, startdate)
-    if enddate != None:
-        ienddate = min(enddate_list)
-        enddate = pd.to_datetime(enddate)
-        enddate = min(ienddate, enddate)
-    if startdate == None:        
-        startdate = max(startdate_list)
-    if enddate == None:
-        enddate = min(enddate_list)
-
-    merged_dataset = merged_dataset.sel(time=slice(startdate, enddate))
-    del reader
-    del data
-    del dataset_list
-    del startdate_list
-    del enddate_list
-    gc.collect()
-    return startdate, enddate, merged_dataset
-
-
-if __name__ == '__main__':
+if __name__ == "__main__":
 
     args = parse_arguments(sys.argv[1:])
 
     loglevel = get_arg(args, "loglevel", "WARNING")
-    logger = log_configure(loglevel, 'CLI multi-model Timeseries ensemble')
-    logger.info("Running Global Time Series diagnostic")
+    logger = log_configure(loglevel, "CLI multi-model Timeseries ensemble")
+    logger.info("Starting Ensemble Time Series diagnostic")
 
-    # Moving to the current directory so that relative paths work
-    abspath = os.path.abspath(__file__)
-    dname = os.path.dirname(abspath)
-    if os.getcwd() != dname:
-        os.chdir(dname)
-        logger.info(f"Changing directory to {dname}")
+    cluster = get_arg(args, "cluster", None)
+    nworkers = get_arg(args, "nworkers", None)
 
-    # Dask distributed cluster
-    nworkers = get_arg(args, 'nworkers', None)
-    if nworkers:
-        cluster = LocalCluster(n_workers=nworkers, threads_per_worker=1)
-        client = Client(cluster)
-        logger.info(f"Running with {nworkers} dask distributed workers.")
+    (
+        client,
+        cluster,
+        private_cluster,
+    ) = open_cluster(nworkers=nworkers, cluster=cluster, loglevel=loglevel)
 
-    # Load configuration file
-    configdir = ConfigPath(loglevel=loglevel).configdir
-    default_config = os.path.join(configdir, "diagnostics", "ensemble",
-                                  "config_timeseries_ensemble.yaml")
-    file = get_arg(args, "config", default_config)
-    logger.info(f"Reading configuration file {file}")
-    config = load_yaml(file)
-    #file = get_arg(args, "config", "config_timeseries_ensemble.yaml")
-    #logger.info(f"Reading configuration file {file}")
-    #config = load_yaml(file)
+    # Load the configuration file and then merge it with the command-line arguments
+    config_dict = load_diagnostic_config(
+        diagnostic="ensemble",
+        config=args.config,
+        default_config="config_timeseries_ensemble.yaml",
+        loglevel=loglevel,
+    )
+    config_dict = merge_config_args(config=config_dict, args=args, loglevel=loglevel)
 
-    variable = config['variable']
-    logger.info(f"Plotting {variable} timeseries")
-    startdate, enddate, plot_std, plot_ensemble_members, ensemble_label, figure_size, ref_label, label_ncol, label_size, pdf_save, units = get_plot_options(
-        config, variable)
+    # Output options
+    outputdir = config_dict["output"].get("outputdir", "./")
+    # rebuild = config_dict['output'].get('rebuild', True)
+    save_netcdf = config_dict["output"].get("save_netcdf", True)
+    save_pdf = config_dict["output"].get("save_pdf", True)
+    save_png = config_dict["output"].get("save_png", True)
+    # dpi = config_dict['output'].get('dpi', 300)
 
-    outputdir = get_arg(args, "outputdir", config["outputdir"])
+    # EnsembleTimeseries diagnostic
+    if "ensemble" in config_dict["diagnostics"]:
+        if config_dict["diagnostics"]["ensemble"]["run"]:
+            logger.info("EnsembleTimeseries module is used.")
 
-    # Monthly model data
-    mon_model = config['models_monthly']
-    mon_model_list = []
-    mon_exp_list = []
-    mon_source_list = []
-    if mon_model != None:
-        mon_model[0]['model'] = get_arg(args, 'model', mon_model[0]['model'])
-        mon_model[0]['exp'] = get_arg(args, 'exp', mon_model[0]['exp'])
-        mon_model[0]['source'] = get_arg(args, 'source', mon_model[0]['source'])
-        for model in mon_model:
-            mon_model_list.append(model['model'])
-            mon_exp_list.append(model['exp'])
-            mon_source_list.append(model['source'])
+            reference = config_dict["references"][0]
+            variable = config_dict["diagnostics"]["ensemble"].get("variable", None)
+            logger.info(f"Variable under consideration: {variable}")
 
-    mon_startdate, mon_enddate, mon_dataset = retrieve_data(
-        variable, models=mon_model_list, exps=mon_exp_list, sources=mon_source_list, startdate=startdate, enddate=enddate)
+            startdate_data = config_dict["diagnostics"]["ensemble"]["params"]["default"].get(
+                "startdate_data", None
+            )
+            enddate_data = config_dict["diagnostics"]["ensemble"]["params"]["default"].get(
+                "enddate_data", None
+            )
+            startdate_ref = config_dict["diagnostics"]["ensemble"]["params"]["default"].get(
+                "startdate_ref", None
+            )
+            enddate_ref = config_dict["diagnostics"]["ensemble"]["params"]["default"].get(
+                "enddate_ref", None
+            )
+            title = config_dict["diagnostics"]["ensemble"]["plot_params"]["default"].get(
+                "title", None
+            )
+            plot_ensemble_members = config_dict["diagnostics"]["ensemble"]["plot_params"][
+                "default"
+            ].get("plot_ensemble_members", True)
 
-    # Annual model data
-    ann_model = config['models_annual']
-    ann_model_list = []
-    ann_exp_list = []
-    ann_source_list = []
-    if ann_model != None:
-        ann_model[0]['model'] = get_arg(args, 'model', ann_model[0]['model'])
-        ann_model[0]['exp'] = get_arg(args, 'exp', ann_model[0]['exp'])
-        ann_model[0]['source'] = get_arg(args, 'source', ann_model[0]['source'])
-        for model in ann_model:
-            ann_model_list.append(model['model'])
-            ann_exp_list.append(model['exp'])
-            ann_source_list.append(model['source'])
+            # Model data
+            # TODO: hourly and daily data
+            models = config_dict["datasets"]
 
-    ann_startdate, ann_enddate, ann_dataset = retrieve_data(
-        variable, models=mon_model_list, exps=mon_exp_list, sources=ann_source_list, startdate=startdate, enddate=enddate)
+            monthly_catalog_list = []
+            monthly_model_list = []
+            monthly_exp_list = []
+            monthly_source_list = []
 
-    # Reference monthly data
-    ref_mon = config['reference_model_monthly']
-    if ref_mon != None:
-        ref_mon_model = ref_mon[0]['model']
-        ref_mon_exp = ref_mon[0]['exp']
-        ref_mon_source = ref_mon[0]['source']
+            annual_catalog_list = []
+            annual_model_list = []
+            annual_exp_list = []
+            annual_source_list = []
 
-    reader = Reader(model=ref_mon_model, exp=ref_mon_exp, source=ref_mon_source,
-                    startdate=mon_startdate, enddate=mon_enddate, areas=False,variable=variable)
-    ref_mon_dataset = reader.retrieve(var=variable)
+            if models is not None:
+                models[0]["catalog"] = get_arg(args, "catalog", models[0]["catalog"])
+                models[0]["model"] = get_arg(args, "model", models[0]["model"])
+                models[0]["exp"] = get_arg(args, "exp", models[0]["exp"])
+                models[0]["source"] = get_arg(args, "source", models[0]["source"])
+                for model in models:
+                    if model["source"] == "aqua-timeseries-monthly":
+                        monthly_catalog_list.append(model["catalog"])
+                        monthly_model_list.append(model["model"])
+                        monthly_exp_list.append(model["exp"])
+                        monthly_source_list.append(model["source"])
+                    if model["source"] == "aqua-timeseries-annual":
+                        annual_catalog_list.append(model["catalog"])
+                        annual_model_list.append(model["model"])
+                        annual_exp_list.append(model["exp"])
+                        annual_source_list.append(model["source"])
 
-    # Reference annual data
-    ref_ann = config['reference_model_annual']
-    if ref_ann != None:
-        ref_ann_model = ref_ann[0]['model']
-        ref_ann_exp = ref_ann[0]['exp']
-        ref_ann_source = ref_ann[0]['source']
+            # Reterive monthly data
+            monthly_dataset = retrieve_merge_ensemble_data(
+                variable=variable,
+                model_list=monthly_model_list,
+                exp_list=monthly_exp_list,
+                source_list=monthly_source_list,
+                startdate=startdate_data,
+                enddate=enddate_data,
+            )
 
-    reader = Reader(model=ref_ann_model, exp=ref_ann_exp, source=ref_ann_source,
-                    startdate=ann_startdate, enddate=ann_enddate, areas=False,variable=variable)
-    ref_ann_dataset = reader.retrieve(var=variable)
+            # Reterieve annual data
+            annual_dataset = retrieve_merge_ensemble_data(
+                variable=variable,
+                model_list=annual_model_list,
+                exp_list=annual_exp_list,
+                source_list=annual_source_list,
+                startdate=startdate_data,
+                enddate=enddate_data,
+            )
 
-    ts = EnsembleTimeseries(var=variable, mon_model_dataset=mon_dataset, ann_model_dataset=ann_dataset,
-                            mon_ref_data=ref_mon_dataset, ann_ref_data=ref_ann_dataset)
-    try:
-        ts.edit_attributes(plot_std=plot_std, plot_ensemble_members=plot_ensemble_members, ensemble_label=ensemble_label,
-                         ref_label=ref_label, figure_size=figure_size, label_ncol=label_ncol, label_size=label_size, pdf_save=pdf_save)
-        ts.run()
-    except Exception as e:
-        logger.error(f'Error plotting {variable} timeseries: {e}')
+            # Reference monthly data
+            ref = config_dict["references"]
+            ref[0]["catalog"] = get_arg(args, "catalog", ref[0]["catalog"])
+            ref[0]["model"] = get_arg(args, "model", ref[0]["model"])
+            ref[0]["exp"] = get_arg(args, "exp", ref[0]["exp"])
+            ref[0]["source"] = get_arg(args, "source", ref[0]["source"])
+            for ref_model in ref:
+                if ref is not None and ref_model["source"] == "aqua-timeseries-monthly":
+                    ref_monthly_catalog = ref_model["catalog"]
+                    ref_monthly_model = ref_model["model"]
+                    ref_monthly_exp = ref_model["exp"]
+                    ref_monthly_source = ref_model["source"]
+                if ref is not None and ref_model["source"] == "aqua-timeseries-annual":
+                    ref_annual_catalog = ref_model["catalog"]
+                    ref_annual_model = ref_model["model"]
+                    ref_annual_exp = ref_model["exp"]
+                    ref_annual_source = ref_model["source"]
+
+            ## Monthly reference data
+            # reader = Reader(
+            #    model=ref_monthly_model,
+            #    exp=ref_monthly_exp,
+            #    source=ref_monthly_source,
+            #    startdate=monthly_startdate,
+            #    enddate=monthly_enddate,
+            #    areas=False,
+            #    variable=variable,
+            # )
+            # monthly_ref_data = reader.retrieve(var=variable)
+
+            ## Annual reference data
+            # reader = Reader(
+            #    model=ref_annual_model,
+            #    exp=ref_annual_exp,
+            #    source=ref_annual_source,
+            #    startdate=annual_startdate,
+            #    enddate=annual_enddate,
+            #    areas=False,
+            #    variable=variable,
+            # )
+            # annual_ref_data = reader.retrieve(var=variable)
+
+            # Monthly reference data
+            ERA5_monthly = "/work/ab0995/a270260/pre_computed_aqua_analysis/IFS-FESOM/historical-1990/global_time_series/netcdf/global_time_series_timeseries_2t_ERA5_era5_mon.nc"
+            monthly_ref_data = xr.open_dataset(
+                ERA5_monthly,
+                drop_variables=[
+                    var for var in xr.open_dataset(ERA5_monthly).data_vars if var != variable
+                ],
+            )
+            # selection ERA5 data on the same time interval -> xarray.DataArray
+            # monthly_ref_data = monthly_ref_data[variable].sel(time=slice(monthly_dataset.time[0], monthly_dataset.time[-1]))
+            monthly_ref_data = monthly_ref_data[variable].sel(
+                time=slice(startdate_ref, enddate_ref)
+            )
+            # Annual reference data
+            ERA5_annual = "/work/ab0995/a270260/pre_computed_aqua_analysis/IFS-FESOM/historical-1990/global_time_series/netcdf/global_time_series_timeseries_2t_ERA5_era5_ann.nc"
+            annual_ref_data = xr.open_dataset(
+                ERA5_annual,
+                drop_variables=[
+                    var for var in xr.open_dataset(ERA5_annual).data_vars if var != variable
+                ],
+            )
+            # selection ERA5 data on the same time interval -> xarray.DataArray
+            # annual_ref_data = annual_ref_data[variable].sel(time=slice(annual_dataset.time[0], annual_dataset.time[-1]))
+            annual_ref_data = annual_ref_data[variable].sel(time=slice(startdate_ref, enddate_ref))
+
+            # Check if we need monthly and annual time variables
+            ts = EnsembleTimeseries(
+                var=variable,
+                monthly_data=monthly_dataset,
+                annual_data=annual_dataset,
+                catalog_list=monthly_catalog_list,
+                model_list=monthly_model_list,
+                exp_list=monthly_exp_list,
+                source_list=monthly_source_list,
+                outputdir=outputdir,
+                loglevel=loglevel,
+            )
+
+            # Compute statistics and save the results as netcdf
+            ts.run()
+
+            # PlotEnsembleTimeseries class
+            plot_arguments = {
+                "var": variable,
+                "catalog_list": monthly_catalog_list,
+                "model_list": monthly_model_list,
+                "exp_list": monthly_exp_list,
+                "source_list": monthly_source_list,
+                "ref_catalog": ref_monthly_catalog,
+                "ref_model": ref_monthly_model,
+                "ref_exp": ref_monthly_exp,
+                "save_pdf": save_pdf,
+                "save_png": save_png,
+                "plot_ensemble_members": plot_ensemble_members,
+                "title": title,
+            }
+
+            ts_plot = PlotEnsembleTimeseries(
+                **plot_arguments,
+                monthly_data=ts.monthly_data,
+                monthly_data_mean=ts.monthly_data_mean,
+                monthly_data_std=ts.monthly_data_std,
+                annual_data=ts.annual_data,
+                annual_data_mean=ts.annual_data_mean,
+                annual_data_std=ts.annual_data_std,
+                ref_monthly_data=monthly_ref_data,
+                ref_annual_data=annual_ref_data,
+                outputdir=outputdir,
+                log_level=loglevel,
+            )
+
+            # plot() function in PlotEnsembleTimeseries class
+            ensemble_plot = ts_plot.plot()
+
+            logger.info(f"Finished Ensemble time series diagnostic for {variable}.")
+
+    close_cluster(
+        client=client, cluster=cluster, private_cluster=private_cluster, loglevel=loglevel
+    )